Constrain space in which upscale bugs can happen
There's a bug where:

- Users can't get their own output grid (sometimes)
- Users can upscale each others outputs (sometimes)

The bug seems to be related to shared state across grids. If a user tries
to upscale across two different outputs at once strange behavior also occurs.

This is likely related to the use of parent_msg in the code, so I've removed it
and replaced it with simple memory of the generation row, which should have more
predictable behavior.